pyproject.toml
[project]
name = "imaging-source-recorder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ['imagingcontrol4>=1.2.0', 'imagingcontrol4pyside6','uvicorn', 'fastapi']

[dependency-groups]
dev = [
    "mypy>=1.15.0",
]

[tool.uv]
package = true

[project.scripts]
imaging-source-recorder-gui = "gui:main_gui"